Update packages & major refactoring of widgets and files description: - update code to comply with new norms in flutter 2.0 - enforce code standards by using pedantic package - separate a collection widgets & widget utility methods in their own files - add & improve code comments - update example code, use new functions or features for demo and use color pick event to update app's accent color on the go
