+++ example/pubspec.yaml
@@ -11,10 +11,10 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.2
 
   #library to store the preference/settings using hive library
-  hive_flutter: ^0.2.1
+  hive_flutter: ^1.0.0
 
 dev_dependencies:
   flutter_test:
