+++ lib/src/utils/utils.dart
@@ -4,12 +4,12 @@ class Utils {
   static Color colorFromString(String colorString) {
     colorString = colorString.replaceFirst('#', '');
     colorString = colorString.length == 6 ? 'ff' + colorString : colorString;
-    int colorHexInt = int.parse(colorString, radix: 16);
+    var colorHexInt = int.parse(colorString, radix: 16);
     return Color(colorHexInt);
   }
 
   static int intFromString(String value) {
-    String alphaString = 'ff000000';
+    var alphaString = 'ff000000';
     return int.parse(value, radix: 16) & int.parse(alphaString, radix: 16);
   }
