+++ pubspec.yaml
@@ -19,13 +19,16 @@ dependencies:
   flutter_material_color_picker: ^1.0.5
 
   #library to access storage directories
-  path_provider: ^1.4.4
+  path_provider: ^2.0.1
 
   #library to use Provider widgets
-  provider: ^4.0.0
+  provider: ^5.0.0
 
   #library to store the preference/settings using shared_preferences library
-  shared_preferences: ^0.5.4+6
+  shared_preferences: ^2.0.4
+
+  # to enforce dart coding standard
+  pedantic: ^1.11.0
 
 dev_dependencies:
   flutter_test:
